pyproject.toml
[tool.poetry]
name = "doraimon"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
python-dotenv = "^1.0.0"
langchain = "^0.0.345"
openai = "^1.3.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
